Support enabling multicast snooping for SecondaryNetwork
SecondaryNetwork allows multicast traffic to be transmitted over a
separate network. While OVS bridges natively support multicast, multicast
snooping is disabled by default. This causes multicast traffic to be
flooded to all ports, which is inefficient in large environments.

This patch introduces a configuration parameter, enableMulticastSnooping,
that enables multicast snooping on the bridge associated with a
SecondaryNetwork. With snooping enabled, the bridge learns multicast
group memberships and forwards multicast traffic only to ports with
interested receivers.

Signed-off-by: Quan Tian <quan.tian@broadcom.com>