ICMP Rule with NodeLatencyMonitor
This PR provides the solution to an edge case with NodeLatencyMonitor, where
the feature does not work if ICMP queries are blocked by default. To fix this,
an iptable rule will be added if NodeLatencyMonitor is enabled, such that ICMP
requests via the Antrea gateway will be allowed.

Fixes issue #6952

Signed-off-by: Dhruv Jain <dhruv.jain@broadcom.com>