Persist container netns with OVS port external IDs
The following agent functions require netns of a container network
interface:
* Create or update a Pod secondary interface.
* Delete a Pod secondary interface and clean up its state.
* Restore the original name of a SR-IOV VF interface after it is removed
from the container netns.
If the container netns info is not persisted, these functions may break
after agent restarts. This commit stores netns in the interface store,
and writes it to an OVS port external ID with a Pod's primary or
a secondary VLAN interface and restores it into the interface store
after agent restarts.

Signed-off-by: Jianjun Shen <shenj@vmware.com>