Fix TestPacketCapture e2e test failure (#7167)
The `invalid-host-public-key` test case is currently failing. The test
case includes a failed SFTP upload, and it needs a custom timeout value
to account for upload retries (and the delay between retries). It is not
clear why it was passing before. It seems that we were right at the
edge, and an update to the Github runner affected the timing and caused
the test to fail.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>