Add TCP flags filter support to PacketCapture (#7070)
For #6864

A new field `flags` is added to the `TCPHeader` specification.
The API provide maximum flexibility, with the ability to provide
multiple filters / matchers (which will be OR'ed), each one with
a value and a mask.

The `antctl packetcapture` command is also updated, with
support for the `tcp_flags` field in the flow specification. The
format for specifying flags to match is the same as for OVS
flows. For example, `tcp_flags=+syn-ack` will match packets
with the SYN flag set and the ACK flag unset, ignoring all the
other flags.

Note that the filter applies identically to packets in both
directions for bi-directional captures.

Signed-off-by: Aryan Bakliwal <aryanbakliwal12345@gmail.com>