Automated cherry pick of #7165: Fix Kind job failures on Github after Docker Engine update #7167: Fix TestPacketCapture e2e test failure (#7167) (#7172)
* Fix Kind job failures on Github after Docker Engine update (#7165)

The version of Docker Engine included in the Github runners was recently
updated from v26 to v28. This has caused consistent CI failures.

* Use `com.docker.network.endpoint.ifname` label when connecting an
  additional network to Kind Nodes to provide an explicit interface name
* Use `gateway_mode_ipv[46]=nat-unprotected` to make sure that Kind
  Nodes on a non-default docker network can acess the K8s apiserver

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>

* Fix TestPacketCapture e2e test failure (#7167)

The `invalid-host-public-key` test case is currently failing. The test
case includes a failed SFTP upload, and it needs a custom timeout value
to account for upload retries (and the delay between retries). It is not
clear why it was passing before. It seems that we were right at the
edge, and an update to the Github runner affected the timing and caused
the test to fail.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>

---------

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>