Remove stale local members in the group cache (#7154)
This change is to resolve the issue that the same receiver may fail to receive
multicast packets after it rejoins the group with encap mode. The issue happens
when the last local member has left the Multicast group, but there exists
receivers located on other Nodes in the cluster.

The issue was introduced because the Multicast controller directly adds the
group into the worker queue but didn't update status in the cache, which makes
the re-join event from the same Pod is ignored.

The fix is to generate IGMP leave message for the stale members even if it is
the last member on the current Node. It also calls checkLastMember in function
`clearStaleGroups` when no local members are left, rather than directly adds the
group into the worker queue. It also removes the field "lastIGMPReport" in
struct GroupMemberStatus.

Fixes #7140

Signed-off-by: Wenying Dong <wenying.dong@broadcom.com>