Fix the flaky failure with Multicast unit test (#7183)
The test case may fail if the check happens when the worker doesn't complete the
functions to sync a group but the no events exist in the queue. The fix is to
add an invalid event into the groupEventCh to ensure the case is completed after all
expected events are processed, and call queue.ShutDownWithDrain to ensure all items
in the queue are processed.

Signed-off-by: Wenying Dong <wenying.dong@broadcom.com>