Refactor declaration of IPFIX IEs for FlowExporter
In #6912, we added support for version skew betwen the FlowExporter
(Agent) and the FlowAggregator. The implementation was based on the
observation that when a new Antrea version introduces additional
Information Elements (IEs), they are always added "after" all existing
IEs. I have recently realized that this is not accurate, due to the
destinationClusterIPv4 / destinationClusterIPv6 always being
conditionally added after all other IEs, including new ones.

To ensure that future new IEs are always added at the end of the list,
we refactor how these IEs are declared. As an added benefit, we now
share the declaration between the FlowExporter and the FlowAggregator,
and the code is overall less verbose.

Note that we should still improve how version skew is handled in the
future, e.g., we should tolerate the removal of deprecated IEs, but this
is hard to do efficiently.

There is also still room for improvement in the FlowAggregator, as we
could consistently use the new style of declaring IEs for the
FA-specific ones.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>