Remove terminated Pods in groupEntityIndex
Once a Pod is in "Succeeded" or "Failed" phase, Kubernetes will not
restart the same Pod. In groupEntityIndex, this should be considered
as a terminal event, and the Pod should be excluded from any Network
Policy computations in appliedTo or address groups. This should also
prevent the antrea agent from reconciling policies for completed Pods
whose interfaces would have already been deleted

Signed-off-by: Dyanngg <dingyang@vmware.com>