Remove unnecessary logs about NPL cleanup
A NPL cleanup log was being generated every time a Pod without NPL
enabled was created, updated, or deleted. This could flood the logs and
potentially confuse users.

This patch removes the code that handles the delete-all-rules case
specifically and logs the event unconditionally even if nothing is done.
It also reduces code redundancy.

Signed-off-by: Quan Tian <quan.tian@broadcom.com>