Periodically sync permanent neighbors to ensure route correctness
In the current implementation, permanent neighbors are installed only once at startup.
However, these neighbors are critical for the functionality of routes managed by antrea-agent.
If they are accidentally deleted, the affected routes will break and remain non-functional
until antrea-agent is restarted.

This commit adds periodic syncing of permanent neighbors to ensure they are reinstalled
automatically if missing, improving the robustness and reliability of route management.

Signed-off-by: Hongliang Liu <hongliang.liu@broadcom.com>