[FlowAggregator] Move aggregation logic from go-ipfix (#7227)
* [FlowAggregator] Move aggregation logic from go-ipfix

Because the aggregation code is Antrea-specific, while the go-ipfix
library should ideally be a generic collection of tools for IPFIX, it
makes more sense to have the code live in this code base.

The code was copied as is, including unit tests, with minimal changes:

- the `IPFIXAggregationProcess` interface was moved from `pkg/ipfix` to
  `pkg/flowaggregator/intermediate` (alongside the migrated code) and
  renamed to `AggregationProcess`. It is used for mocking in unit tests.
- the `AggregationProcess` struct (in the migrated code) was renamed to
  `aggregationProcess`.

* Use virtual clock for some of the aggregation testing

Some unit tests were failing on Windows (Github Actions) because the
system clock was not advancing as expected (lower clock resolution on
Windows runner?).

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>