Increase the default packet-in rate limit to 5000 (#7243)
Benchmark tests show that handling 10,000 packet-in events for DNS
interception and 10,000 events for NetworkPolicy logging work well. The
patch increases the default packet-in rate limit to prevent applications
that generate high volumes of DNS queries from being impacted by rate
limiting.

Although 10,000 is supported, this patch sets the default value to 5000,
which should be sufficient for most use cases. It also stops advising
users not to raise the limit.

Signed-off-by: Quan Tian <quan.tian@broadcom.com>