Simplify sanitizeExtractPath (used when extracting archive) (#7234)
Github code analysis is currently reporting (incorrectly) that the code
is vulnerable to a "Zip Slip" attack. This may be because of how we
handle the case where the destination directory is "." (the working
directory). We simplify the code (and eliminate the special case) by
using filepath.IsLocal, which was introduced in Go 1.20. Note that an
alternative would have been to make the paths absolute, but this does
not work well when using an arbitrary afero.FS. When using
filepath.IsLocal, absolute paths will be rejected (an alternative could
be to first strip the leading /, but this support is not required for our
use case). We also add unit tests to validate the function.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>