[FlowAggregator] Use Protobuf message internally (#7253)
When an IPFIX record is received by the IPFIX collector, it is
converted to a Protobuf message. This provides stronger decoupling
between the Flow Exporter and the Flow Aggregator, and also simplifies
the logic in the Flow Aggregator (in particular correlation /
aggregation). The Protobuf message is the only data structure used in
the Flow Aggregator, and the exporters are in charge of converting it
to the right format based on the destination.

With this change, it becomes much easier to support version skew
between the Flow Exporter and the Flow Aggregator.

A follow-up change will switch to using the Protobuf format directly
(with a gRPC stream) between the Flow Exporter and the Flow Aggregator.
We will preserve backwards compatibility for a few versions (the Flow
Aggregator will keep accepting IPFIX records).

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>