Automated cherry pick of #7183: Fix the flaky failure with Multicast unit test (#7183) #7202: Fix flaky failure in unit test case (#7262)
* Fix the flaky failure with Multicast unit test (#7183)

The test case may fail if the check happens when the worker doesn't complete the
functions to sync a group but the no events exist in the queue. The fix is to
add an invalid event into the groupEventCh to ensure the case is completed after all
expected events are processed, and call queue.ShutDownWithDrain to ensure all items
in the queue are processed.

Signed-off-by: Wenying Dong <wenying.dong@broadcom.com>

* Fix flaky failure in unit test case TestClearStaleGroups (#7202)

The root cause of the flaky failure is one case is designed with multiple local
members of a stale Multicast group. According to latest implementation, the
mcast_controller will send IGMP leave event standing for all the local members
one by one. And there exists two separate goroutines, eventHandler and worker.
The worker may take different actions with the OpenFlow group according to the
cached GroupMemberStatus in the runtime, e.g., 1. if there exists other local
members (the event handler does not remove other stale local members) then the
worker will re-install the OpenFlow group to update the local members, 2. if all
local members are removed from the cache, the worker will un-install the
OpenFlow group. Finally the OpenFlow group is removed.

To remove the flaky cases, this change use "MaxTimes" to mock the call with
ofClient.InstallMulticastGroup on such Mutlicast groups.

Signed-off-by: Wenying Dong <wenying.dong@broadcom.com>

---------

Signed-off-by: Wenying Dong <wenying.dong@broadcom.com>