Fix e2e test TestAntreaNodeNetworkPolicy failing in noEncap/hybrid mode
For NodeNetworkPolicy, there are two hostNetwork Pods and non-hostNetwork Pod
created for testing.

- Pod Ax, non-hostNetwork on Node A
- Pod Ay, hostNetwork on Node A
- Pod Bz, hostNetwork on Node B

In encap mode, to support UDP/SCTP NodeNetworkPolicy tests, the gateway IPs are
used as the destination IP for hostNetwork Pods. The reason is that when using the
Node external IPs as destination IPs when probing, the UDP/SCTP reply traffic from
hostNetwork Pods will choose a source IP address based on the routing decision or
outgoing interface, which means that the local Antrea gateway IPs will be chosen
as the source IP address.

When using a Node external IP, the mismatch of source and destination IPs causes
UDP/SCTP probing failure.

- UDP/SCTP Request: - Pod Ax -> Node A external IP -> Pod Ay
- UDP/SCTP Reply:   - Pod Ay -> Node A gateway IP  -> Pod Ax

The probing can work with a gateway IP.

- UDP/SCTP Request: - Pod Ax -> Node A gateway IP  -> Pod Ay
- UDP/SCTP Reply:   - Pod Ay -> Node A gateway IP  -> Pod Ax

Other successful cases in encap mode:

- Pod Ax -> Node B gateway IP  -> Pod Bz
- Pod Ay -> Node B gateway IP  -> Pod Bz
- Pod Bz -> Node A gateway IP  -> Pod Ay

However, in noEncap mode, these tests fail because reply traffic is sent via
the transport interface (with the external IP):

- Pod Ax -> Node B gateway IP  -> Pod Bz, fails
- Pod Ay -> Node B gateway IP  -> Pod Bz, fails
- Pod Bz -> Node A gateway IP  -> Pod Ay, fails

Only this passes. The destination is the gateway IP and the reply traffic is
also sent from the gateway interface.

- Pod Ax -> Node A gateway IP  -> Pod Ay, succeeds

Only use the Node gateway IP as a destination when:

- The destination Pod is hostNetwork
- The source Pod is non-hostNetwork
- Both Pods are on the same Node

Signed-off-by: Hongliang Liu <hongliang.liu@broadcom.com>