[FlowAggregator] Ability to export K8s UIDs in the IPFIX exporter
The IPFIX exporter in the Flow Aggregator now has the ability to export
the UIDs of K8s objects (source Pod, source Node, destination Pod,
destination Node, etc.) using the appropriate Antrea IEs. These IEs all
have type OctetArray, with a fixed length of 16 bytes, and are named
with the "UUID" suffix, e.g., sourcePodUUID.

In addition to exporting K8s UIDs, it is also possible to choose not to
export the corresponding K8s names, in order to reduce the size of the
IPFIX data records.

Note that other Flow Aggregator exporters / sinks, are not affected.

We try to populate UIDs in the Flow Protobuf message from the Flow
Exporter when it is convenient. Otherwise, we populate the missing ones
from the Flow Aggregator. In addition to the existing PodStore, the Flow
Aggregator now also needs to run a NodeStore and a ServiceStore. In
order to avoid code duplication, we introduce a generic ObjectStore
which is embedded by the object-specific stores. UID fields also
participate in correlation when using Aggregate mode.

Note that when setting flowCollector.includeK8sUIDs to true in the Flow
Aggregator configuration, the version of the Agent should be at least
v2.4, or some UIDs may be missing (i.e., set to "zero").

E2E testing is a bit limited: we only test this change in Proxy mode,
and we do not yet do a comprehensive validation of all the new IEs.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>