Initialize SRIOV interfaces via NetworkStatus when agent restarts (#7240)
When the agent restarts, current interfaceStore for SecondaryNetwork will notrecover the assigned SRIOV
interfaces for containers because of no OVS bridge to check, so we leverage the NetworkStatus info
in the Pod annotation to recover the interface info.

This change would help to ensure the interface name restoration for SR-IOV devices after agent restarts,
and also fix the issue that IP is not released when a Pod withsecondary SR-IOV nic is deleted after agent restarts.

Signed-off-by: Lan Luo <lan.luo@broadcom.com>