Fix flaky UT of TestPodControllerRun (#7276)
primaryInterfaceStore is expected to be empty before a new Pod is added.
Otherwise, `reconcileSecondaryInterfaces` may restore the cniCache
unexpectly and treat a Pod add event as update and skip reconciling.

The reason is main test routine and `go podController.Run(stopCh)`
execution sequence is not guaranteed. So remove unused test codes to
fix flaky UT of TestPodControllerRun.

Signed-off-by: Lan Luo <lan.luo@broadcom.com>