Use Protobuf / gRPC between FlowExporter and FlowAggregator (#7264)
The FlowExportService gRPC Service, which defines a client-streaming
RPC, is introduced as an alternative to IPFIX. We make it the new
default for exporting flows from the FlowExporter (Agent) to the
FlowAggregator. The exported flows are defined using the Protobuf
message introduced in #7253.

The FlowAggregator will always run the gRPC collector. It will also run
the IPFIX collector by default (tls protocol), but that collector can be
disabled by setting aggregatorTransportProtocol to "none".
This means that a FlowAggregator can support older Agents (IPFIX export)
and newer Agents (gRPC export) simultaneously.

Some fields were updated in the Protobuf message in a non-backwards
compatible way, which is fine because until now the message was only
used internally within the FlowAggregator process, and was never
serialized. All the "UID" fields now have type string (instead of
bytes).

The test-e2e-flow-visibility Github workflow job will now run for both
IPFIX and gRPC.

This change is a bit more verbose than I initially planned, because I
did some refactoring in pkg/agent/flowexporter to have a more elegant
package structure, while avoiding cyclic dependencies.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>