Update github.com/ClickHouse/clickhouse-go/v2 to v2.35.0
This dependency has not been updated in a long time. One of its
dependencies (github.com/ClickHouse/ch-go) is currently affected by a
CVE, and updating github.com/ClickHouse/clickhouse-go/v2 gets rid of the
security alert.

We cannot update to the latest released version (v2.37.2), as it results
in a build error on Windows.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>