Fix e2e test TestAntreaNodeNetworkPolicy failing in noEncap/hybrid mode (#7267)
\In encap mode, to support UDP/SCTP NodeNetworkPolicy tests, the gateway IPs are
used as the destination IP for hostNetwork Pods. The reason is that when using the
Node external IPs as destination IPs when probing, the UDP/SCTP reply traffic from
hostNetwork Pods will choose a source IP address based on the routing decision or
outgoing interface, which means that the local Antrea gateway IPs will be chosen
as the source IP address.

However, in noEncap mode, these tests fail because reply traffic is sent via
the transport interface (with the external IP):

In order for the tests to support both cases, only use the Node gateway IP
as a destination when:

- The destination Pod is hostNetwork
- The source Pod is non-hostNetwork
- Both Pods are on the same Node

Signed-off-by: Hongliang Liu <hongliang.liu@broadcom.com>