Fix Github status reporting for build job (#7296)
The "build" job includes an execution matrix, which is not expanded when
the job is skipped, causing some required status checks to stay in the
Pending state, preventing PRs from being merged.
This is a known issue, and my proposed workaround is to introduce a
separate "build-status" job. This new job should replace "build" as the
required status check for PRs, as it will always report a status. Of
course, "build-status" is guaranteed to fail if "build" fails or is
cancelled.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>