Add resource requests for Windows container in Antrea deployment (#7254)
Based on performance testing on a 4C8G Windows node running 100 pods,
the antrea-agent container showed an average CPU usage of 30m and memory
usage of 50MB, while the OVS container consumed 17m CPU and 23MB memory.

To account for potential burst scenarios and ensure runtime stability,
memory requests are set to 100MB and CPU requests are set to 200m (except
for the install-cni initContainer). The CPU requests match the ones for the
Agent on Linux.

Signed-off-by: Shuyang Xin <gavinx@vmware.com>