Remove trailing whitespace from default manifests (#7311) (#7312)
When the flowExporter.protocolFilter Helm value is nil, we need to make
sure that we don't add a trailing whitespace in the rendered
antrea-agent.conf YAML. Otherwise this leads to formatting issues (no
pretty-printing) when displaying or editing the antrea-config ConfigMap.

This is a known "issue":
https://github.com/kubernetes/kubernetes/issues/75446.

We already have at least one other configuration value,
egress.snatFullyRandomPorts, for which we can have a nil value and that
is handled in the same way.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>