[FlowExporter] Ignore conntrack connections denied by policy
We filter out connections from the Antrea CT zone that were committed
but later dropped by a policy rule. Examples include a Pod-to-Service
connection which is dropped by an egress rule, and an intra-Node
Pod-to-Pod connection which is dropped by an ingress rule. In order to
achieve this, we introduce a new CT mark (`ConnAllowedCTMark = 1 << 8`),
which is applied to connections in the ConntrackCommit table, after
egress & ingress policy enforcement - an exception is hairpin & SNAT
connections for which the mark needs to be applied earlier. Connections
which do not have the mark will be collected by the FlowExporter using
the "deny connection" mechanism (based on OVS PacketIn instead of
conntrack). By ignoring the conntrack connection, we avoid duplicates
across both connection stores and possible issues in the FlowAggregator
caused by duplicate flow exports. This change will resolve a test
failure issue caused by the FlowAggregator processing the "conntrack"
version of the flow before the "deny" version of the flow (now the
"conntrack" version of the flow is no longer exported).

We also enhance the FlowAggregator e2e tests to validate that egress
policy information is reported correctly when a connection is denied by
an ingress policy. Note that previously this wasn't always the case (for
intra-Node flows).

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>