Add missing Run calls for nodeStore / serviceStore (#7343) (#7359)
* [FlowAggregator] Add missing Run calls for nodeStore / serviceStore

Without a call to Run, there is no worker to consume items from the
deletion workqueue, and the size of the store will keep increasing until
the process runs out-of-memory. We add the missing Run calls, which are
now in cmd/flow-aggregator.

Additionally, we wait for the nodeStore and serviceStore to be synced
before we start processing flows in the FlowAggregator. This is
consistent with what we already do for the podStore, and helps avoid
failed store lookups.



* Address review comment



---------

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>
Co-authored-by: Antonin Bas <antonin.bas@broadcom.com>