Refine AntreaProxy framework to eliminate duplicated processing in dual-stack
This commit introduces a wrapper `ProxyServer`, which holds shared objects used by IPv4
and IPv6 proxier. Common processing, such as running `serviceConfig`, `endpointSliceConfig`,
`nodeConfig` and feature gates, are moved to `ProxyServer` to avoid duplicated processing
when AntreaProxy is running in dual-stack.

Additionally, this commit fixes the Prometheus metrics defined for AntreaProxy.

This commit is also the foundation of introducing a single-instance `ProxyHealthServer` later,
which should remain unique even in dual-stack.

Signed-off-by: Hongliang Liu <hongliang.liu@broadcom.com>