[PacketCapture] Support arbitrary source or destination (#7215)
This change makes the `source` and `destination` fields optional in the
PacketCapture spec.
The only requirement is that one of `source.pod` or `destination.pod` be
set.
If the source is omitted, traffic from any source is captured.
If the destination is omitted, traffic from any destination is captured.

Fixes #6976

Signed-off-by: Harsh Gupta <notsoharsh13@gmail.com>