[FlowExporter] Ignore conntrack connections denied by policy (#7361)
We filter out connections from the Antrea CT zone that were committed
but later dropped by a policy rule. Examples include a Pod-to-Service
connection which is dropped by an egress rule, and an intra-Node
Pod-to-Pod connection which is dropped by an ingress rule. In order to
achieve this, we rely on `ConnSourceCTMarkField` (bits 0-3 of the CT
mark), and we ensure that this CT mark field remains unset until after
both egress & ingress policy enforcement (until the ConntrackCommit
table) - an exception is hairpin & SNAT connections for which the field
needs to be set earlier, but such connections cannot be dropped by
ingress policy rules. Connections which do not have the field set will
be collected by the FlowExporter using the "deny connection" mechanism
(based on OVS PacketIn instead of conntrack). By ignoring the conntrack
connection, we avoid duplicates across both connection stores and
possible issues in the FlowAggregator caused by duplicate flow
exports. This change will resolve a test failure issue caused by the
FlowAggregator processing the "conntrack" version of the flow before the
"deny" version of the flow (now the "conntrack" version of the flow is
no longer exported).

We also enhance the FlowAggregator e2e tests to validate that egress
policy information is reported correctly when a connection is denied by
an ingress policy. Note that previously this wasn't always the case (for
intra-Node flows).

Co-authored-by: Quan Tian <quan.tian@broadcom.com>

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>