Update dev VERSION automatically using Github workflow (#7418)
For now, this is done with a separate job (and a separate PR) in the
post_release_actions.yml workflow. We could unify the VERSION update PR
and the CHANGELOG update PR in the future.

If we need to bump the *major* version number instead of the *minor*
version number, the PR can be discarded (or overridden afterwards).

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>