More robust way to extract source Node IP from encapsulated IGMP messages (#7282) (#7379)
In hybrid or encap mode, IGMP reports between Nodes could be sent via tunnels using packet-out,
with the Node transport IP as the source for encapsulated packets. For remote Nodes:

- The underlay packet (carrying the encapsulated IGMP report) uses the Node transport IP as
  its source.
- The remote Node extracts this source IP of the underlay packet to determine `tun_dst`
  and install OVS flow-based tunnel flows.
- The installed OVS flow-based tunnel flows ensures that the multicast traffic is correctly
  forwarded over tunnel.

In hybrid or encap mode, however, the underlay packets may be SNATed before reaching the remote
Node. This means:

- The underlay packet source IP could be altered and no longer reflect the Node transport IP.
- Relying on the underlay source IP would cause incorrect OVS flow-based tunnel flows to be
  installed.

The Node transport IP is used as the source IP of the IGMP report packets. When sending the
packet via tunnel, the source IP of the underlay packet that encapsulating the IGMP packets
is the Node transport IP. On the remote Node receiving the packets, the source IP of the
underlay packet are used as the tun_dst to install flow-based tunnel flows, which ensure
that the multicast traffic can be forwarded to remote Node via tunnel properly.

To handle hybrid or encap mode properly, we are using a more robust way to extract the source
Node IP from encapsulated IGMP messages:

- Extract the source IP from the overlay IGMP packet, not the underlay one.
- The overlay packet’s source stays as the Node transport IP, even if underlay SNAT occurs.

Signed-off-by: Hongliang Liu <hongliang.liu@broadcom.com>