Add network-policy-delay flag to antctl check installation (#7427)
Instead of using a hard-coded 2s value.
While I am not able to reproduce the issue locally (even without a sleep), the
egress-deny-all-connectivity and ingress-deny-all-connectivity tests sometimes
fail in CI, presumably because the policy has not been realized by all Agents
yet. So to help prevent the issue, we use a 5s delay in CI, using the new flag.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>