Fix flaky TestGoBGPLifecycle integration test
The major change is that we are forcing the connection mode for each
side of the BGP session: one side active and one side passive. This
eliminates connection "collisions" and back-offs: by making Server2
always the active connector and Server1 always the passive listener, we
prevent both sides from trying to connect simultaneously. In particular,
when we update the peer as a test step (which requires re-establishing
the session), the connection modes remain the same and reconnection
succeeds immediately. This is better than increasing polling timeouts to
very large values.

We also improve the test by using a different localhost address
(127.0.0.1 and 127.0.0.2) for each server, and matching the router ID to
the listen address. This requires explicitly setting the "local address"
used to connect to the peer, for each server.

Finally, we also improve our logging wrapper for gobgp: we include the
router ID in each log message and we make sure that the source location
logged is the correct one (gobgp code, not log wrapper function) by
going one call frame up the stack.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>