Add conntrack_poll_cycle_duration_seconds metric
To keep track of how long it takes to poll conntrack and update the
connection store. This could help troubleshoot issues. We also log the
duration, using the existing verbosity (`V(2)`).

Note that unlike some other existing metrics, we use seconds as the
unit, not milliseconds. Using seconds seems to be best practice for
this type of Prometheus metrics.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>