[FlowExporter] Avoid invalid NetworkPolicy data in records
To avoid missing or invalid NetworkPolicy data in flow records, we
increase the time interval during which the mapping from flow ID to
NetworkPolicy rule is available in the idAllocator after rule
deletion. During that time, the flow ID cannot be reused.

We also avoid resolving NetworkPolicy metadata for connections which
started prior to the Agent starting and the initial NetworkPolicy flows
being installed. After an Agent restart, the flow ID assignment is
expected to change, and we do not want to include invalid NetworkPolicy
data for these connections (better to omit it altogether). We only make
this change for the "conntrack" connection store. The "deny" connection
store is not as sensitive to this because it reacts to PacketIn messages
and not conntrack entries (which can exist for a while and persist
during an Agent restart).

For #7438

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>