[Flaky unit test] Fix TestCluster_RunClusterEvents (#7467)
This is a follow-up to #7448.
The change in behavor of assert.Eventually has caused to the unit test
to become flaky. This change improves how the result of egress selection
is validated, and resolves the flakiness issue.
It is expected that the test function can be greatly improved when we
upgrade to Go 1.25 and we can start using the new testing/synctest
package. Upgrading to Go 1.25 is not immediately possible because of
some dependency issues.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>