[FlowExporter] Use netlink zone filtering when dumping ct flows (#7504)
The Antrea FlowExporter dumps flows by zone when polling
conntrack. Until now, filtering was done by the Antrea code itself, in
userspace. This wastes memory and a lot of CPU time (netlink decoding).

Support for the CTA_ZONE attribute for dump and flush requests was added
to the kernel in 6.8. We update the ti-mo/conntrack library to v0.6.0,
which leverages this support, and we invoke DumpFilter with the correct
zone filter.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>