Clean up stale secondary IPs in IPPool when Node restarts with invalid OVSDB
When a Kubernetes Node reboots and the OVSDB file is not properly restored,
the primary and secondary OVS ports are lost. Pods may temporarily
appear as "unknown" and later get recreated with new container IDs.

The cleanup of secondary interfaces during agent restart relies on the
InterfaceStores being initialized with existing OVS ports. If the OVSDB file
is missing, the primary InterfaceStore is empty, which prevents the secondary
InterfaceStore from initializing correctly. Consequently, secondary IPs assigned
to old Pods (with previous container IDs) are not deleted, leaving stale entries
in the IPPool.

This change adds logic to periodically verify all IPPool allocations and release
any secondary IPs associated with non-existing container IDs. This also covers
the case where a Pod is recreated with the same name on another Node and has any
secondary IPs.

Signed-off-by: Lan Luo <lan.luo@broadcom.com>