Clean up stale secondary IPs in IPPool when Node restarts with invalid OVSDB
When a Kubernetes Node reboots and the OVSDB file is not properly
restored, the primary InterfaceStore cannot be recovered. Since the
cleanup of existing secondary interfaces during agent restart depends
on the primary InterfaceStore, stale secondary IP addresses remain in
the IPPool. These stale IPs cannot be reused.

To address this, add a one-time cleanup step that runs on agent restart
to remove stale IP addresses from the IPPool.

Signed-off-by: Lan Luo <lan.luo@broadcom.com>