Handle missing Pod IP and Pod IP changes (#7512)
The existing code in waitForRulesInitialization was assuming that
status.podIP was always set. This led to the installation of
incorrect datapath rules (iptables rules on Linux Nodes), which were
never corrected later on by the controller. We address this by skipping
rule initialization for Pods with missing IPs, and by clearing their NPL
annotation. We also add validation to make sure that all required fields
are present in each NPL annotation entry, otherwise we also skip
initialization.

Additionally, also update the Pod handler to handle IP resets /
changes. There are 2 edge cases in which this can be observed:
  * a Pod Sandbox can be recreated (e.g., after a Node restart). In this
    case, we observe that the IP is lost and that a new IP is later
    assigned to the Pod.
  * a Pod is deleted and immediately recreated with the same name but a
    different IP address (on the same Node): because the NPLController
    uses a workqueue, the handler can observe a single Pod "update" and
    the same Pod name is associated with a different IP.

Finally, we move unit tests for the NPLController to the
`pkg/agent/nodeportlocal/k8s` package, where they more naturally belong.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>