[Flaky unit test] Fix TestSyncEgress on Windows
The TestSyncEgress unit test was failing frequently because of an order
mismatch between expected Events and actual Events. The actual Events
could not be sorted correctly because the LastTimestamp was sometimes
the same for the AssignedEgress Event and for the UnassignedEgress
Event. This seems to be caused by a "resolution" issue with the
monotonic clock on the Windows Github runner. To avoid the issue
altogether, we use a custom Event handler in the test, which means that
we no longer need to rely on the search API for Events. Our Event
handler ensures that Events are always correctly ordered based on the
order in which they are received. Because the unit test calls the
`syncEgress` method directly, and does not rely on worker goroutines,
this order is always deterministic.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>