[Flaky unit test] Fix TestServiceLifecycle (#7531)
The test would sometimes fail because it was possible for the
EndpointSlice creation to lead to an extra workqueue item being
processed.

We also remove the context parameter from processNextWorkItem, as it was
misleading: workqueue operations don't have context parameters and
queue.Get() can block indefinitely regardless of the context. The only
signal that processNextWorkItem can use is the workqueue being shut down.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>