Accelerate Pod-to-Pod networking in noEncap/hybrid mode (#7324)
* Accelerate Pod-to-Pod networking in noEncap/hybrid mode

This patch introduces nftables flowtable support to accelerate Pod-to-Pod
traffic in the Node's host network when the traffic mode is noEncap or hybrid.
By offloading flows to the nftables flowtable, most processing hooks in the
Node's host network stack are bypassed, reducing overhead.

A new nftables table antrea is added, which includes:

- Sets of peer Pod CIDRs (IPv4 or IPv6)
- A flowtable bound to antrea-gw0 and the transport interface.
- Forwarding rules that match Pod-to-Pod traffic between local and
  remote Pod CIDRs, and insert flows into the flowtable for acceleration.

Signed-off-by: Hongliang Liu <hongliang.liu@broadcom.com>