Improve stale IP recycling in AntreaIPAM controller
When a Node is deleted, the CNI DEL command may not be invoked for Pods
running on that Node. As a result, antrea-agent cannot release their IPs,
leaving stale entries in the IPPool. To mitigate this, increase the IP
recycle interval from 10 minutes to 1 minute.

Additionally, add a benchmark test for the IP recycle function and refine
its implementation based on benchmark results to reduce memory allocations
and improve cleanup performance.

Signed-off-by: Lan Luo <lan.luo@broadcom.com>