[FlowAggregator] Only generate certificate if needed
This addresses an issue where having multiple replicas of
flow-aggregator will cause it to clobber each other's
certificates. We now only try to create it if it is invalid
(does not exist/expired). The root CA certificate and key
is shared amongst all instances. Each replica is reponsible
for generating their own server certificates. Client certs
and the CA ConfigMap will be createed/updated by a single
replica.

Signed-off-by: Andrew Su <andrew.su@broadcom.com>